refactor(viz): restructure Python files into subfolders
Move html_generator.py and layout_estimator.py into new html/ subpackage,
move instructions.py into renderer/, and delete dead code (traversal.py,
coordinates.py).
