+++ src/hypergraph/viz/html/generator.py
