+++ src/hypergraph/viz/renderer/instructions.py
